% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_subgroups.R
\name{error_subgroups}
\alias{error_subgroups}
\title{Divide observations into groups based on OOB predicted error of original residuals}
\usage{
error_subgroups(Y, X, preds, variables = NULL, num_groups = 10)
}
\arguments{
\item{Y}{a vector of responses}

\item{X}{a matrix or data frame of covariates}

\item{preds}{a vector of predicted values}

\item{variables}{a vector of variable names in X,
May also be NULL, in which case all variables in X are used.}

\item{num_groups}{the number of subgroups to create}
}
\value{
A vector specifying subgroup assignment, of the same length as Y.
Smaller values indicate the predicted error was smaller.
}
\description{
\code{error_subgroups} runs a random forest to predict the absolute value of
the residuals calculated from \code{Y} and \code{preds} using all variables
specified in \code{variables}. Observations are divided into
subgroups based on the values of the OOB predicted error from this random
forest.
}
