% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mean_imputation}
\alias{mean_imputation}
\title{Mean Imputation}
\usage{
mean_imputation(df)
}
\arguments{
\item{df}{data frame to modify}
}
\value{
\code{mean_imputation} returns the original data frame, modified so that all missing values are
have been replaced with the mean of the column. All columns with missing values
now have an additional corresponding column with \verb{_mis} tagged onto
the original column name. The new columns will have a 1 in any row where there
was originally a missing value in the corresponding column and a 0 otherwise.
}
\description{
Replaces all missing values with the mean of the column and creates new
columns to capture what values were initially missing.
}
