% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mode_calc}
\alias{mode_calc}
\title{Calculating the mode for a specified column of a data frame}
\usage{
mode_calc(df, col)
}
\arguments{
\item{df}{a data frame to analyze}

\item{col}{the name of a column in \code{df} to calculate the mode for}
}
\value{
The value that appears most frequently in the \code{var} column of \code{df}
}
\description{
Calculating the mode for a specified column of a data frame
}
