% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_subgroups.R
\name{numeric_subgroups}
\alias{numeric_subgroups}
\title{Divide observations into groups based on a single numeric variable}
\usage{
numeric_subgroups(Y, X, grouping_col, preds, max_groups = 10)
}
\arguments{
\item{Y}{vector of responses}

\item{X}{a matrix or dataframe of covariates}

\item{grouping_col}{A column in X, used to divide observations into subgroups}

\item{preds}{vector of predictions}

\item{max_groups}{number of groups to create}
}
\value{
A vector of subgroup assignment
}
\description{
Divide observations into groups based on a single numeric variable
}
