% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_size.R
\name{samp_size}
\alias{samp_size}
\title{Calculate necessary sample size}
\usage{
samp_size(sigma2, effect_size, alpha = 0.05, beta = 0.2, type = "two.sided")
}
\arguments{
\item{sigma2}{An estimate of the variance of the population}

\item{effect_size}{desired minimal detectible difference}

\item{alpha}{desired Type II error rate}

\item{beta}{desired Type II error rate}

\item{type}{A character string specifying the type of test. Must be either
"two.sided" (default) or "one.sided"}
}
\value{
The number of observations needed for the overall experiment
}
\description{
\code{samp_size} calculates the needed sample size for the entire
experiment to achieve the stated power, type I error, and effect size,
assuming equally sized groups.
}
\details{
Sample size is calculated using the following equation:
\deqn{n = 4 \sigma^2 \frac{(Z_{\alpha/2} + Z_{\beta})^2}{\Delta^2}}
where \eqn{\Delta} is the effect size, \eqn{Z_{\alpha/2}} is \eqn{1-\alpha/2}
quantile value from a normal distribution, and \eqn{Z_{\beta}} is the \eqn{beta}
quantile value from a normal distribution. If \code{type = "one.sided"}, then
\eqn{Z_{\alpha/2}} is replaced with \eqn{Z_{\alpha}}.
}
