% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Create interactive Shiny app to calculate sample sizes using reloop}
\usage{
run_app(...)
}
\arguments{
\item{...}{}
}
\description{
Create interactive Shiny app to calculate sample sizes using reloop
}
\details{
\code{run_app} uses other \code{reloop.samp} functions to display a Shiny app that walks users
through the process of defining subgroups and calculating the sample size that would be
necessary if the RCT population resembled each subgroup. Sample sizes for each subgroup
without using ReLOOP are also displayed. Additionally, users can explore diagnostics
of the initial random forest results and can compare each subgroup to the rest of the
remnant for any covariate.
}
